requirements.txt
streamlit
neo4j
pandas
python-dotenv
streamlit-agraph